the next approach to make the dongle flashing via bluetooth easier..
this time as java program. Device discovery works, also the download
of the flashfile from a URL.
Not implemented so far:
- tested only under linux
- ymodem (still trying to find some source...)
- proxy handling for downloads from inside coorperate networks

Usage:
"make all" to compile
"make run" to run

Actual the software just downloads the firmware zip file and interprets
the content, it also scans for dongles, reads their ID string and starts the
bootloader, but just resets the device from there - ymodem comes, as soon
some source is found or written...

Actual the software works just on the console. Feel free to write a simple GUI
for it ;-)