fundamental change in serial_dxm, webUI now with dynamic element generation
a very basic initialisation sequence was located in serial_dxm, which forced to load that file
and initialize the whole serial stuff, even if not needed at all.

this sequence is now moved into lua_utils, so this file now always need to be included (if not already
done at all)

the webUI now supports dynamic elements generations, which makes it basically compatible to all
old scripts, as long as no advanced functions as file download, remote Connect etc. is needed

Also timer and update flags are not correctly supported yet - this stuff comes next..