Function-freeze Level reached
This is the first, reference version of the final function content of OOBD. As of todays
knowledge, this software now contains everything which has been found as necessary to have.
After having this, we now move over to spread the functionality into the other platforms,
do some bug hunting and clean up, but more functions should not be expected. Also the old
GUI will removed in parallel.

Thanks to all team members, namely Mike, Wolfgang(s), Josef, Uli, Alex (and even Ekkehard in his role as
permanent detractor :-) ) that we all together made it and reached this goal.

So let us clean up before the first visitors come :-)