"Lost connection" added
Up to now it was quite unclear if an answer from the dongle was empty as such or if there is no
connection to the dongle anymore it all.

Now the function echoWrite(stringToSend) reports back, if the sented string did came back from
the dongle as echo, or if even this feedback is missing, which indicates a broken connection
to the dongle itself.

As echoWrite() shall not be used in user applications, this new feature is implemented now in the
uds_lua- Library, which reports a "Lost Connection" as new return value in case the dongle does
not echo the command.