canlauncher ok for experimental use
canlauncher is a program launcher to start other programs like candump etc., which
do rely on the availability of the cansocket CAN device and which stops on bus loss.

canlauncher monitors these bus loss and restart the program, if needed, so e.g. with

  canlauncher can0 candump can0 > canlog.txt

you'll get a continious log file even if the bus gets lost sometime in between

for more parameters, start canlauncher without arguments