As far as I can see with my limited testing capabilities: App CarDTC supports USD Ping mode for quick bus scans
Technical details:
The possible module adresses and names per carmaker are defined in cardtcs.js, so new brands and names
can be add easily just by editing that file.

The user can then select, which busses he expects (only standard, standard+low speed or 2 high speeds)
Then, when press read, on these buses a 190102 (get Nr. of DTCs) is send on broadcast ID 7DF
From each first frame from each module, which answers, byte 5-7 are taken as number of DTCs.
If the module ID is contained in the carddtc.js table, its color in the list changes to green or yellow,
depending on the number of DTCs (guess, for what green is for...)
When then selecting a single module, it DTC list can be viewed and deleted.