Performance improvement: Posix firmware changed from single char transmission to whole blocks
Because of the carry over routines when changed from serial to tcpip telnet connection, each char was
probably transmitted as single tcp packet, which might have caused a performance leak. Now an internal
buffer is filled first and sent as once at the end of the command execution.

Please note that the new routine flushSerial() needs to be added as empty routine also to serial
implementations to avoid linker errors